Add wrapup_event and limit heartbeat to heartbeat
Why:

The heartbeat is tied with information which is not related to
"heartbeat" concept. Also, heartbeat should be quite frequent to be
useful but frequent updates of entire info and captured out can be very
costly.

How:

All updates but heartbeat are moved to wrapup event. Also both wrapup
event and heartbeat are turned into specific updates on the row rather
full updates on the entire row.